overhaul path updates to user mistakes

+++ src/routes.js
@@ -14,24 +14,26 @@ export const routes = [
             if (id != undefined && name != undefined && content != undefined) { return res.end('Para pesquisas com filtros, utilize um (e apenas um) filtro de pesquisa!') }
             if (id != undefined && name != undefined || id != undefined && content != undefined || name != undefined && content != undefined) { return res.end('Para pesquisas com filtros, utilize um (e apenas um) filtro de pesquisa!') }
 
-            if (!id == undefined) {
+
+            if (id != undefined) {
                 const tasks = database.select('tasks', id ? { id: id } : [])
 
                 if (tasks.length == 0) return res.end('Nada foi encontrado')
                 else return res.end(JSON.stringify(tasks))
             }
-            if (!name == undefined) {
+            if (name != undefined) {
                 const tasks = database.select('tasks', name ? { name: name } : [])
 
                 if (tasks.length == 0) return res.end('Nada foi encontrado')
                 else return res.end(JSON.stringify(tasks))
             }
-            if (!content == undefined) {
+            if (content != undefined) {
                 const tasks = database.select('tasks', content ? { content: content } : [])
 
                 if (tasks.length == 0) return res.end('Nada foi encontrado')
                 else return res.end(JSON.stringify(tasks))
             }
+
 
             if (id == undefined && name == undefined && content == undefined) {
                 const tasks = database.select('tasks', id ? { id: id } : [])
@@ -48,7 +50,7 @@ export const routes = [
             const { name, content } = req.body
 
             if (typeof (content) !== "string" || typeof (name) !== "string") { return res.end("Valores inválidos") }
-            if (content === " " || name === " ") { return res.end("Valores inválidos") }
+            if (content === " " || name === " ") { return res.end("Os valores estão em branco, por favor, digite eles.") }
 
             const tasks = {
                 id: randomUUID(),
@@ -69,10 +71,14 @@ export const routes = [
         path: buildRoutePath('/tasks/delete'),
         handler: (req, res) => {
             const { id } = req.query
-
+        
+            if(!id){return res.end("Por favor, insira o id da tarefa a ser deletada nos parâmetros!")}
+            
+            const tasks = database.select('tasks', id ? { id: id } : [])
+            if (tasks.length == 0) return res.end('A Tarefa a ser deletado na tabela não existe')
             database.delete('tasks', id)
 
-            return res.writeHead(204).end()
+            return res.end('Tarefa Deletado!')
         },
     },
     {
@@ -82,13 +88,18 @@ export const routes = [
             const { id } = req.query
             const { name, content, isClosed } = req.body
 
-            if (typeof (content) !== "string" && content !== undefined || typeof (name) !== "string" && name !== undefined || typeof (isClosed) !== 'boolean' && isClosed !== undefined) { return res.end("Valores inválidos") }
-            if (content === " " || name === " ") { return res.end("Valores inválidos") }
+
+            if(!id){return res.end("Por favor, insira o id da tarefa a ser editada nos parâmetros!")}
+            const tasks = database.select('tasks', id ? { id: id } : [])
+            if (tasks.length == 0) return res.end('A tarefa a ser editada na tabela não existe')
+            if (!content && !name && !isClosed){ return res.end("Não insira valores vazios") }
+
+            if(typeof (content) !== "string" && content !== undefined){ return res.end("Valores inválidos para content") }
+            if(typeof (name) !== "string" && name !== undefined ){return res.end("Valores inválidos para name")}
+            if(typeof (isClosed) !== 'boolean' && isClosed !== undefined){return res.end("Valores inválidos para isClosed")}
 
             if (isClosed == undefined) {
                 const tasks = database.update('tasks', id, {
-                    name,
-                    content,
                     isClosed: false,
                     updatedAt: new Date(),
 
@@ -96,7 +107,16 @@ export const routes = [
                 return res.end(JSON.stringify(tasks))
             }
 
-            if (name == undefined) {
+
+            if((content == undefined || content == " ") && (name == undefined || name == " ")){
+                const tasks = database.update('tasks', id, {
+                    isClosed,
+                    updatedAt: new Date(),
+
+                })
+                return res.end(JSON.stringify(tasks))
+            }
+            if (name == undefined || name == " ") {
                 const tasks = database.update('tasks', id, {
                     content,
                     isClosed,
@@ -105,7 +125,7 @@ export const routes = [
                 })
                 return res.end(JSON.stringify(tasks))
             }
-            if (content == undefined) {
+            if (content == undefined || content == " " ) {
                 const tasks = database.update('tasks', id, {
                     name,
                     isClosed,
